Display filtered recipes on URL change
the useLocation hook by React Router allow to listen URL changes and to access
the current URL search parameters. The useeffect calls, every time there is a
change in the URL, an external function that fetch the recipes from the API.
Returned recipes are store in state and each create a Preview component.